Moving consts to reviews.ts, refactor const into JSX
