+++ site/src/helpers/reviews.ts
@@ -26,3 +26,22 @@ export function getReviewHeadingName(
     return `${instructor?.name}`;
   }
 }
+
+export function formatQuarter(quarter: string): string {
+  const [year, term] = quarter.split(' ');
+  const termMap: Record<string, string> = {
+    Fall: 'F',
+    Winter: 'W',
+    Spring: 'Sp',
+    Summer: 'Su',
+  };
+  return `${termMap[term] ?? term}${year.slice(-2)}`;
+}
+
+export function displayReviewDate(date: string | Date): string {
+  return new Date(date).toLocaleString('default', {
+    year: 'numeric',
+    month: '2-digit',
+    day: '2-digit',
+  });
+}
